材质过滤。

+++ ShadowEditor.Web/src/editor/bottom/MaterialPanel.js
@@ -1,5 +1,6 @@
 import UI from '../../ui/UI';
 import Ajax from '../../utils/Ajax';
+import EditWindow from '../window/EditWindow';
 import MaterialsSerializer from '../../serialization/material/MaterialsSerializer';
 
 /**
@@ -12,7 +13,6 @@ function MaterialPanel(options) {
 
     this.firstShow = true;
 
-    this.keywords = '';
     this.data = [];
 };
 
@@ -57,6 +57,9 @@ MaterialPanel.prototype.renderUI = function () {
             },
             children: [{
                 xtype: 'div',
+                style: {
+                    display: 'flex'
+                },
                 children: [{
                     xtype: 'searchfield',
                     id: 'search',
@@ -72,16 +75,9 @@ MaterialPanel.prototype.renderUI = function () {
                 },
                 children: [{
                     xtype: 'category',
-                    options: {
-                        category1: '分类1',
-                        category2: '分类2',
-                        category3: '分类3',
-                        category4: '分类4',
-                        category5: '分类5',
-                        category6: '分类6',
-                        category7: '分类7',
-                        category8: '分类8',
-                    }
+                    id: 'category',
+                    scope: this.id,
+                    onChange: this.onSearch.bind(this)
                 }]
             }]
         }, {
@@ -98,7 +94,7 @@ MaterialPanel.prototype.renderUI = function () {
                 scope: this.id,
                 style: {
                     width: '100%',
-                    height: '100%',
+                    maxHeight: '100%',
                 },
                 onClick: this.onClick.bind(this)
             }]
@@ -109,29 +105,57 @@ MaterialPanel.prototype.renderUI = function () {
 };
 
 MaterialPanel.prototype.update = function () {
-    var server = this.app.options.server;
+    this.updateCategory();
+    this.updateList();
+};
 
-    this.keywords = '';
+MaterialPanel.prototype.updateCategory = function () {
+    var category = UI.get('category', this.id);
+    category.clear();
 
-    Ajax.getJson(`${server}/api/Material/List`, obj => {
-        this.data = obj.Data;
-        this.onSearch('');
+    Ajax.getJson(`/api/Category/List?type=Material`, obj => {
+        category.options = {};
+        obj.Data.forEach(n => {
+            category.options[n.ID] = n.Name;
+        });
+        category.render();
     });
 };
 
-MaterialPanel.prototype.onSearch = function (name) {
-    if (name.trim() === '') {
-        this.renderList(this.data);
-        return;
+MaterialPanel.prototype.updateList = function () {
+    var search = UI.get('search', this.id);
+
+    Ajax.getJson(`/api/Material/List`, obj => {
+        this.data = obj.Data;
+        search.setValue('');
+        this.onSearch();
+    });
+};
+
+MaterialPanel.prototype.onSearch = function () {
+    var search = UI.get('search', this.id);
+    var category = UI.get('category', this.id);
+
+    var name = search.getValue();
+    var categories = category.getValue();
+
+    var list = this.data;
+
+    if (name.trim() !== '') {
+        name = name.toLowerCase();
+
+        list = list.filter(n => {
+            return n.Name.indexOf(name) > -1 ||
+                n.FirstPinYin.indexOf(name) > -1 ||
+                n.TotalPinYin.indexOf(name) > -1;
+        });
     }
 
-    name = name.toLowerCase();
-
-    var list = this.data.filter(n => {
-        return n.Name.indexOf(name) > -1 ||
-            n.FirstPinYin.indexOf(name) > -1 ||
-            n.TotalPinYin.indexOf(name) > -1;
-    });
+    if (categories.length > 0) {
+        list = list.filter(n => {
+            return categories.indexOf(n.CategoryID) > -1;
+        });
+    }
 
     this.renderList(list);
 };
@@ -161,15 +185,23 @@ MaterialPanel.prototype.onClick = function (event, index, btn, control) {
     var data = control.children[index].data;
 
     if (btn === 'edit') {
-        this.onEdit(data);
+        if (typeof (this.onEdit) === 'function') {
+            this.onEdit(data);
+        }
     } else if (btn === 'delete') {
-        this.onDelete(data);
+        if (typeof (this.onDelete) === 'function') {
+            this.onDelete(data);
+        }
     } else {
-        this.onSelect(data);
+        if (typeof (this.onClick) === 'function') {
+            this.onSelectMaterial(data);
+        }
     }
 };
 
-MaterialPanel.prototype.onSelect = function (data) {
+// ------------------------------------- 选择 ------------------------------------
+
+MaterialPanel.prototype.onSelectMaterial = function (data) {
     Ajax.get(`/api/Material/Get?ID=${data.ID}`, result => {
         var obj = JSON.parse(result);
         if (obj.Code === 200) {
@@ -179,12 +211,39 @@ MaterialPanel.prototype.onSelect = function (data) {
     });
 };
 
+
+// ------------------------------- 编辑 ---------------------------------------
+
 MaterialPanel.prototype.onEdit = function (data) {
-    UI.msg('编辑材质');
+    if (this.editWindow === undefined) {
+        this.editWindow = new EditWindow({
+            app: this.app,
+            parent: document.body,
+            type: 'Material',
+            typeName: '材质',
+            saveUrl: '/api/Material/Edit',
+            callback: this.update.bind(this)
+        });
+        this.editWindow.render();
+    }
+    this.editWindow.setData(data);
+    this.editWindow.show();
 };
 
+// -------------------------------- 删除 ----------------------------------------
+
 MaterialPanel.prototype.onDelete = function (data) {
-    UI.msg('删除材质');
+    UI.confirm('询问', `是否删除${data.Name}？`, (event, btn) => {
+        if (btn === 'ok') {
+            Ajax.post(`/api/Material/Delete?ID=${data.ID}`, json => {
+                var obj = JSON.parse(json);
+                if (obj.Code === 200) {
+                    this.update();
+                }
+                UI.msg(obj.Msg);
+            });
+        }
+    });
 };
 
 export default MaterialPanel;
